pyproject.toml
[tool.poetry]
name = "langchain-langsmith-chatbot"
version = "0.1.0"
description = "An experimental chatbot to demonstrate AI engineering skills."
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "chatbot", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
langchain = "^1.2.10"
langgraph = "^1.0.9"
langchain-openai = "^1.1.10"
tavily-python = "^0.3.3"
python-dotenv = "^1.0.1"
ipykernel = "^6.29.4"
langchainhub = "^0.1.16"
langchain-community = "^0.4.1"
langchain-tavily = "^0.2.17"
langchain-google-genai = "^4.2.1"
fastapi = "^0.129.1"
uvicorn = "^0.41.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
